pyproject.toml
[project]
name = "google-a2a-agents-samples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[a2a,eval]==1.13.0",
    "pyyaml>=6.0.2",
    "pydantic>=2.11.7",
    "a2a-sdk>=0.3.5",
    "langchain-community==0.3.27",
    "stackapi==0.3.1",
]
